Refactor TimberModule methods
