+++ modules/TimberModule.php
@@ -5,6 +5,8 @@ namespace Sitchco\Modules;
 use Sitchco\Framework\Module;
 
 use Sitchco\Support\DateTime;
+use Sitchco\Utils\ArrayUtil;
+use Sitchco\Utils\Block;
 use Timber\Timber;
 use Sitchco\Utils\TimberUtil;
 use WP_Block;
@@ -16,8 +18,6 @@ use Traversable;
  */
 class TimberModule extends Module
 {
-    protected static array $blockMetadataCache = [];
-
     public function init(): void
     {
         if (class_exists('Timber\Timber')) {
@@ -33,7 +33,7 @@ class TimberModule extends Module
                 'callable' => [TimberUtil::class, 'includeWithContext'],
             ];
             $functions['inner_blocks'] = [
-                'callable' => [self::class, 'renderInnerBlocksTag'],
+                'callable' => [Block::class, 'renderInnerBlocksTag'],
                 'is_safe' => ['html'],
             ];
             $functions['render_block'] = [
@@ -82,49 +82,18 @@ class TimberModule extends Module
         int $post_id = 0,
         WP_Block $wp_block = null,
     ): void {
-        $context = Timber::context();
-        $context['post'] = $post_id ? Timber::get_post($post_id) : Timber::get_post();
-        $context['fields'] = get_fields();
-        $context['is_preview'] = $is_preview;
-        $context['content'] = $content;
-
-        if ($wp_block instanceof WP_Block) {
-            $childBlocks = iterator_to_array($wp_block->inner_blocks);
-            $block['innerBlocks'] = array_map(static fn(WP_Block $child) => $child->parsed_block, $childBlocks);
-        } elseif (!isset($block['innerBlocks']) && $content) {
-            $block['innerBlocks'] = parse_blocks($content);
-        }
-        $context['block'] = $block;
+        $context = static::setupContext(...func_get_args());
 
         // Parent theme context inclusion
         $context = static::loadBlockContext($context, $block['path']);
-        // Child theme context inclusion
-        $path_parts = explode('/modules/', $block['path']);
-        if (isset($path_parts[1])) {
-            $child_path = get_stylesheet_directory() . '/modules/' . $path_parts[1];
-            $context = static::loadBlockContext($context, $child_path);
-        }
 
-        $metadata = static::getBlockMetadata($block['path'] ?? '');
-        if ($metadata) {
-            $innerBlocksDefaults = $metadata['innerBlocksConfig'] ?? [];
-            if (isset($metadata['allowedBlocks']) && !isset($innerBlocksDefaults['allowedBlocks'])) {
-                $innerBlocksDefaults['allowedBlocks'] = $metadata['allowedBlocks'];
-            }
-            if ($innerBlocksDefaults) {
-                $existingInnerBlocksConfig = $context['innerBlocksConfig'] ?? [];
-                if (is_array($existingInnerBlocksConfig)) {
-                    $innerBlocksDefaults = array_merge($innerBlocksDefaults, $existingInnerBlocksConfig);
-                }
-                $context['innerBlocksConfig'] = $innerBlocksDefaults;
-                if (!isset($context['allowedBlocks']) && isset($innerBlocksDefaults['allowedBlocks'])) {
-                    $context['allowedBlocks'] = $innerBlocksDefaults['allowedBlocks'];
-                }
-                if (!isset($context['innerBlocksTemplate']) && isset($innerBlocksDefaults['template'])) {
-                    $context['innerBlocksTemplate'] = $innerBlocksDefaults['template'];
-                }
-            }
-        }
+        // Child theme context inclusion
+        $relative = Block::relativeBlockPath($block['path']);
+        $child_path = get_stylesheet_directory() . '/modules/' . $relative;
+        $context = static::loadBlockContext($context, $child_path);
+
+        $metadata = Block::getBlockMetadata($block['path'] ?? '');
+        $context = static::mergeMetadataContext($context, $metadata);
         $context = static::normalizeInnerBlocksContext($context);
 
         // Auto-inject helper variables for templates
@@ -141,57 +110,50 @@ class TimberModule extends Module
         echo TimberUtil::compileWithContext($template_path, $context, "block/$blockName");
     }
 
-    public static function renderInnerBlocksTag(array $config = []): string
+    protected static function setupContext(
+        array $block,
+        string $content = '',
+        bool $is_preview = false,
+        int $post_id = 0,
+        WP_Block $wp_block = null,
+    ): array {
+        $context = Timber::context();
+        $context['post'] = $post_id ? Timber::get_post($post_id) : Timber::get_post();
+        $context['fields'] = get_fields();
+        $context['is_preview'] = $is_preview;
+        $context['content'] = $content;
+
+        if ($wp_block instanceof WP_Block) {
+            $childBlocks = iterator_to_array($wp_block->inner_blocks);
+            $block['innerBlocks'] = array_map(static fn(WP_Block $child) => $child->parsed_block, $childBlocks);
+        } elseif (!isset($block['innerBlocks']) && $content) {
+            $block['innerBlocks'] = parse_blocks($content);
+        }
+        $context['block'] = $block;
+        return $context;
+    }
+
+    protected static function mergeMetadataContext(array $context, array $metadata): array
     {
-        if (isset($config['allowed']) && !isset($config['allowedBlocks'])) {
-            $config['allowedBlocks'] = $config['allowed'];
+        $innerBlocksDefaults = $metadata['innerBlocksConfig'] ?? [];
+        if (isset($metadata['allowedBlocks']) && !isset($innerBlocksDefaults['allowedBlocks'])) {
+            $innerBlocksDefaults['allowedBlocks'] = $metadata['allowedBlocks'];
         }
-        if (isset($config['template_lock']) && !isset($config['templateLock'])) {
-            $config['templateLock'] = $config['template_lock'];
+        if (empty($innerBlocksDefaults)) {
+            return $context;
         }
-
-        $attributeStrategies = [
-            'allowedBlocks' => 'json',
-            'template' => 'json',
-            'templateLock' => 'string',
-            'orientation' => 'string',
-            'renderAppender' => 'bool',
-            'templateInsertUpdatesSelection' => 'bool',
-            'layout' => 'json',
-            'align' => 'string',
-            'className' => 'string',
-        ];
-
-        $attributes = [];
-        foreach ($attributeStrategies as $key => $strategy) {
-            if (!array_key_exists($key, $config)) {
-                continue;
-            }
-            $formattedValue = static::formatInnerBlocksAttribute($config[$key], $strategy);
-            if ($formattedValue !== null) {
-                $attributes[$key] = $formattedValue;
-            }
+        $existingInnerBlocksConfig = $context['innerBlocksConfig'] ?? [];
+        if (is_array($existingInnerBlocksConfig)) {
+            $innerBlocksDefaults = array_merge($innerBlocksDefaults, $existingInnerBlocksConfig);
         }
-
-        if (isset($config['attributes']) && is_array($config['attributes'])) {
-            foreach ($config['attributes'] as $key => $value) {
-                $formattedValue = static::formatInnerBlocksAttribute($value);
-                if ($formattedValue !== null) {
-                    $attributes[$key] = $formattedValue;
-                }
-            }
+        $context['innerBlocksConfig'] = $innerBlocksDefaults;
+        if (!isset($context['allowedBlocks']) && isset($innerBlocksDefaults['allowedBlocks'])) {
+            $context['allowedBlocks'] = $innerBlocksDefaults['allowedBlocks'];
         }
-
-        if (!$attributes) {
-            return '<InnerBlocks />';
+        if (!isset($context['innerBlocksTemplate']) && isset($innerBlocksDefaults['template'])) {
+            $context['innerBlocksTemplate'] = $innerBlocksDefaults['template'];
         }
-
-        $parts = [];
-        foreach ($attributes as $key => $value) {
-            $parts[] = sprintf('%s="%s"', esc_attr($key), esc_attr($value));
-        }
-
-        return '<InnerBlocks ' . implode(' ', $parts) . ' />';
+        return $context;
     }
 
     protected static function loadBlockContext(array $context, $path): array
@@ -211,12 +173,12 @@ class TimberModule extends Module
             $config = [];
         }
 
-        $allowedBlocks = static::normalizeIterable($context['allowedBlocks'] ?? null);
+        $allowedBlocks = ArrayUtil::normalizeIterable($context['allowedBlocks'] ?? null);
         if ($allowedBlocks !== null) {
             $config['allowedBlocks'] = $allowedBlocks;
         }
 
-        $template = static::normalizeIterable($context['innerBlocksTemplate'] ?? null, preserveKeys: false);
+        $template = ArrayUtil::normalizeIterable($context['innerBlocksTemplate'] ?? null, preserveKeys: false);
         if ($template !== null) {
             $config['template'] = $template;
         }
@@ -226,77 +188,5 @@ class TimberModule extends Module
         }
 
         return $context;
-    }
-
-    protected static function normalizeIterable(mixed $value, bool $preserveKeys = true): ?array
-    {
-        if ($value === null) {
-            return null;
-        }
-        if (is_array($value)) {
-            return $preserveKeys ? $value : array_values($value);
-        }
-        if ($value instanceof Traversable) {
-            return iterator_to_array($value, $preserveKeys);
-        }
-        return null;
-    }
-
-    protected static function getBlockMetadata(string $path): array
-    {
-        if (!$path) {
-            return [];
-        }
-        if (array_key_exists($path, static::$blockMetadataCache)) {
-            return static::$blockMetadataCache[$path];
-        }
-        $metadataFile = trailingslashit($path) . 'block.json';
-        if (!is_readable($metadataFile)) {
-            static::$blockMetadataCache[$path] = [];
-            return static::$blockMetadataCache[$path];
-        }
-        $raw = file_get_contents($metadataFile);
-        if (!is_string($raw)) {
-            static::$blockMetadataCache[$path] = [];
-            return static::$blockMetadataCache[$path];
-        }
-        $decoded = json_decode($raw, true);
-        static::$blockMetadataCache[$path] = is_array($decoded) ? $decoded : [];
-        return static::$blockMetadataCache[$path];
-    }
-
-    public static function resetBlockMetadataCache(): void
-    {
-        static::$blockMetadataCache = [];
-    }
-
-    protected static function formatInnerBlocksAttribute(mixed $value, string $strategy = 'auto'): ?string
-    {
-        if ($value === null) {
-            return null;
-        }
-
-        switch ($strategy) {
-            case 'json':
-                $encoded = wp_json_encode($value);
-                if ($encoded === false) {
-                    return null;
-                }
-                return $encoded;
-            case 'bool':
-                return $value ? 'true' : 'false';
-            case 'string':
-                return (string) $value;
-            case 'auto':
-            default:
-                if (is_bool($value)) {
-                    return $value ? 'true' : 'false';
-                }
-                if (is_array($value) || is_object($value)) {
-                    $encoded = wp_json_encode($value);
-                    return $encoded === false ? null : $encoded;
-                }
-                return (string) $value;
-        }
     }
 }
